+++ spb/backends/bokeh/bokeh.py
@@ -386,7 +386,10 @@ class BokehBackend(Plot):
 
         xlims, ylims = [], []
         for r in self.renderers:
-            if r.series.is_interactive:
+            if (
+                r.series.is_interactive
+                or hasattr(r.series, "_interactive_app_controls")
+            ):
                 r.update(params)
                 if hasattr(r, "_xlims"):
                     xlims.extend(r._xlims)